# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BHUVISION \u2014 Agentic Earth Intelligence",
    "description": "An Interactive Vision-Language Assistant for Multimodal Remote Sensing Image Analysis through Text Queries (SIH26167, ISRO). Powered by 9 autonomous specialist agents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/locations/search": {
      "get": {
        "tags": [
          "Locations & Geospatial"
        ],
        "summary": "Search Locations",
        "description": "Search any location on Earth by name and return WGS84 bounding box and MGRS coordinates.",
        "operationId": "search_locations_api_locations_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Location name, city, landmark, or region",
              "title": "Q"
            },
            "description": "Location name, city, landmark, or region"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 6,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationResult"
                  },
                  "title": "Response Search Locations Api Locations Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/nasa-tile": {
      "get": {
        "tags": [
          "NASA Earthdata & GIBS"
        ],
        "summary": "Proxy Nasa Gibs Tile",
        "description": "Proxy authenticated NASA GIBS WMTS tiles with NASA Earthdata Bearer token.\n\nSupports all major MODIS Terra/Aqua and VIIRS SNPP datasets. Returns JPEG tile.\nFalls back to unauthenticated public endpoint if token is not configured.",
        "operationId": "proxy_nasa_gibs_tile_api_nasa_tile_get",
        "parameters": [
          {
            "name": "layer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "NASA GIBS layer name",
              "title": "Layer"
            },
            "description": "NASA GIBS layer name"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date in YYYY-MM-DD format",
              "title": "Date"
            },
            "description": "Date in YYYY-MM-DD format"
          },
          {
            "name": "z",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 15,
              "minimum": 0,
              "description": "Zoom level",
              "title": "Z"
            },
            "description": "Zoom level"
          },
          {
            "name": "y",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Tile Y coordinate",
              "title": "Y"
            },
            "description": "Tile Y coordinate"
          },
          {
            "name": "x",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Tile X coordinate",
              "title": "X"
            },
            "description": "Tile X coordinate"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationResult": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bbox"
          },
          "mgrs": {
            "type": "string",
            "title": "Mgrs"
          },
          "elevation_m": {
            "type": "integer",
            "title": "Elevation M"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "name",
          "display_name",
          "lat",
          "lon",
          "bbox",
          "mgrs",
          "elevation_m",
          "category",
          "country",
          "confidence"
        ],
        "title": "LocationResult",
        "description": "Structured location search result."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}